HobbyDetailActivity에서 HobbyActivity로 코드 이동
