+++ app/src/main/res/layout/activity_hobby_detail.xml
@@ -6,127 +6,6 @@
     android:layout_height="match_parent"
     tools:context=".ui.hobby_detail.HobbyDetailActivity">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/constraintLayout"
-        android:layout_width="match_parent"
-        android:layout_height="45dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
 
-        <TextView
-            android:id="@+id/hobbyName"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="9.5dp"
-            android:layout_marginBottom="13dp"
-            android:fontFamily="@font/notosanscjkkrmedium"
-            android:includeFontPadding="false"
-            android:lineSpacingExtra="7.5sp"
-            android:textColor="#000000"
-            android:textSize="15sp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/writeBtn"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="9.5dp"
-            android:layout_marginEnd="15dp"
-            android:layout_marginBottom="13dp"
-            android:fontFamily="@font/notosanscjkkrmedium"
-            android:includeFontPadding="false"
-            android:lineSpacingExtra="7.5sp"
-            android:text="글쓰기"
-            android:textColor="#9d3cc8"
-            android:textSize="15sp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/hobbySearchLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/constraintLayout">
-
-        <TextView
-            android:id="@+id/tagShowTextView"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="17.5dp"
-            android:layout_marginTop="15dp"
-            android:layout_marginEnd="24dp"
-            android:layout_marginBottom="6.8dp"
-            android:ellipsize="end"
-            android:fontFamily="@font/notosanscjkkrmedium"
-            android:includeFontPadding="false"
-            android:lineSpacingExtra="6.5sp"
-            android:maxLines="1"
-            android:textAlignment="textStart"
-            android:textColor="#000000"
-            android:textSize="14sp"
-            app:layout_constraintBottom_toBottomOf="@id/view"
-            app:layout_constraintEnd_toStartOf="@+id/dial"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <LinearLayout
-            android:id="@+id/dial"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="15dp"
-            android:layout_marginEnd="17.5dp"
-            android:layout_marginBottom="6.8dp"
-            android:gravity="center"
-            android:orientation="horizontal"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <TextView
-                android:id="@+id/textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginRight="5.1dp"
-                android:fontFamily="@font/notosanscjkkrmedium"
-                android:includeFontPadding="false"
-                android:lineSpacingExtra="6.5sp"
-                android:text="필터"
-                android:textColor="#9d3cc8"
-                android:textSize="14sp" />
-
-            <ImageView
-                android:id="@+id/imageView"
-                android:layout_width="17.5dp"
-                android:layout_height="17.5dp"
-                app:srcCompat="@drawable/icon_filter" />
-
-            <TextView
-                android:id="@+id/textView4"
-                android:layout_width="0dp"
-                android:layout_height="0dp"
-                android:layout_weight="1" />
-        </LinearLayout>
-
-        <View
-            android:id="@+id/view"
-            android:layout_width="0dp"
-            android:layout_height="1dp"
-            android:layout_marginStart="15.3dp"
-            android:layout_marginEnd="14.8dp"
-            android:background="#9d3cc8"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="1.0"
-            app:layout_constraintStart_toStartOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
